introduce a hardcoded slack into motion detection

+++ app/src/main/java/com/mobileproj/dynascope/FloatArrayUtil.kt
@@ -38,9 +38,9 @@ fun FloatArray.stats(): String {
 
 fun FloatArray.innerproduct(other: FloatArray) = zip(other).map { it.first * it.second }.sum()
 
-fun FloatArray.autocorr(n: Long): Float {
+fun FloatArray.autocorr(n: Int): Float {
     // drop first n measurements
-    val shifted = this.copyOfRange((n * 3).toInt(), this.size)
+    val shifted = this.copyOfRange(n * 3, this.size)
     val slice = this.sliceArray(0..shifted.size)
     val innerProduct = slice.innerproduct(shifted)
     val shiftedNorm = sqrt(shifted.innerproduct(shifted))
